♻️ (syncing) sync library and items seperately
This commit refactors the code to sync the library and items separately in the syncing service. It introduces two new functions, ﻿syncItems and ﻿syncCollections, which handle syncing the items and collections respectively. This separation improves code organization and makes it easier to understand and maintain the syncing logic.

BREAKING CHANGE: 💥️ functions